validate: don't turn all NS records insecure on non-existent DS
Currently this lists all NS records in the auth_selected from
the same query UID and marks them as insecure.

It should only mark NS records matching the DS record as insecure,
as parent NS records may be in the same response and the absence
of DS doesn't say anything about them.